+++ lib/src/model/modal_theme.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
-import 'dart:ui';
 
 /// Configure modal style
 @immutable